requirements.txt
requests
beautifulsoup4
rapidfuzz
pillow
customtkinter
